README.md
This Python project calculates a person's age in various time units such as years, months, days, hours, and seconds. It takes the date of birth as input and provides an accurate breakdown of the person's age. The project is designed to be simple and easy to use.

Run on Google Colab:

1. Open Google Colab.
2.Copy & paste the Python code into a new Colab notebook.
3.Run the notebook to get the results